[IMP] web: log errors when missing dependencies in lazy loaded bundles
Before this commit, lazily loading a bundle would not display any errors
at all in case of missing dependencies. This makes it quite difficult to
understand a problem when it happens.  This commit makes sure that we
call the function doing the check. Note that it means that missing
dependencies are now "hard" errors: the web client is unusable once it
happens.

closes odoo/odoo#110663

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>